proper hooks are now passed
also can clear canvas now.
However images fill a div which i am unclear which is it. So the images stretches.
Also clearing the canvas also clears the image as the image is now attached to the canvas. However this will make it easier to save the image as the image is now a part of the canvas